move MemberInfoExtensions to it's own file
